+++ retrocache/src/test/java/com/andiag/retrocache/CallbackTest.java
@@ -11,6 +11,8 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.Executor;
 import java.util.concurrent.TimeUnit;
@@ -23,8 +25,10 @@ import retrofit2.Callback;
 import retrofit2.Response;
 import retrofit2.Retrofit;
 import retrofit2.http.GET;
+import retrofit2.http.Path;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.fail;
@@ -149,6 +153,74 @@ public class CallbackTest {
         });
     }
 
+    @Test
+    public void testConcurrentAccess() {
+        List<Thread> threads = new ArrayList<>();
+
+        /* Set up the mock webserver */
+        MockResponse resp = new MockResponse().setBody("VERY_BASIC_BODY");
+
+        for (int i = 0; i < 10; i++) {
+            threads.add(createWorkerThread(mRetrofit.create(DemoService.class), resp, new Callback<String>() {
+                @Override
+                public void onResponse(Call<String> call, Response<String> response) {
+                    assertEquals(response.body(), "VERY_BASIC_BODY");
+                    assertNotNull(mMockCachingSystem.get(Utils.urlToKey(call.request().url())));
+                }
+
+                @Override
+                public void onFailure(Call<String> call, Throwable t) {
+                    fail("Failure executing the request: " + t.getMessage());
+                }
+            }));
+        }
+        for (Thread thread : threads) {
+            thread.start();
+        }
+
+        for (Thread thread : threads) {
+            try {
+                thread.join();
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+        }
+        assertFalse("test", false);
+        mMockCachingSystem.clear();
+    }
+
+    private Thread createWorkerThread(final DemoService service, final MockResponse resp, final Callback<String> callback) {
+        return new Thread() {
+            int sMaxNumberOfRun = 1000;
+            final CachedCall<String> call = service.getHomeById(getId());
+
+            @Override
+            public void run() {
+                try {
+                    int numberOfRun = 0;
+                    while (numberOfRun++ < sMaxNumberOfRun) {
+                        Thread.sleep((long) (Math.random() * 2));
+                        mServer.enqueue(resp.clone());
+                        double choice = Math.random();
+                        if (choice < 0.4) {
+                            call.clone().enqueue(callback);
+                        } else if (choice < 0.7) {
+                            call.clone().refresh(callback);
+                        } else if (choice < 1) {
+                            call.clone().remove();
+                            assertNull(mMockCachingSystem.get(Utils.urlToKey(call.request().url())));
+                        } else {
+                            // do nothing
+                        }
+                    }
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+            }
+        };
+    }
+
+
     private static class MainThreadExecutor implements Executor {
         @Override
         public void execute(@NonNull Runnable command) {
@@ -159,5 +231,8 @@ public class CallbackTest {
     interface DemoService {
         @GET("/")
         CachedCall<String> getHome();
+
+        @GET("/{id}")
+        CachedCall<String> getHomeById(@Path("id") long id);
     }
 }