+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# NEXT
+    - Method `getCachingSystem()` added to `CachedCallFactory` class
+
 # 1.1.1
     - Added `clone()` method to `CachedCall` interface to avoid extra casting.
     - Implemented new tests and fixed old ones.